requirements.txt
starlette
fastapi
uvicorn
pydantic
pyyaml
requests
apscheduler
sqlalchemy
psycopg2-binary